% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstar.R
\name{ash2}
\alias{ash2}
\title{Replicates ash1 but without print statement}
\usage{
ash2(bins, m = 5, kopt = c(2, 2))
}
\arguments{
\item{bins}{number of bins}

\item{kopt}{}
}
\value{

}
\description{
Replicates ash1 but without print statement
}
