% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstar.R
\name{create_kde}
\alias{create_kde}
\title{Create kde give a set of draws}
\usage{
create_kde(x, method = "bkde", bandwidth = NULL, ...)
}
\arguments{
\item{x}{a set of univariate draws}

\item{method}{a kde method (either bkde or ash1)}

\item{bandwidth}{bandwidth of kde}

\item{...}{other arguments passed to kde method}
}
\value{
a function which when evaluated returns estimated density at a point
}
\description{
Create kde give a set of draws
}
