% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstar.R
\name{dstar}
\alias{dstar}
\title{Calculates the d* measure of information gain from after seeing the data in
Bayesian inference}
\usage{
dstar(
  x1,
  x2,
  f1 = NULL,
  f2 = NULL,
  method = "bkde",
  training_percent = 0.7,
  bandwidth_1 = NULL,
  bandwidth_2 = NULL,
  ...
)
}
\arguments{
\item{x1}{a set of univariate draws of a parameter from the prior}

\item{x2}{a set of univariate draws of a parameter from the posterior}

\item{f1}{if}

\item{...}{}
}
\value{

}
\description{
Calculates the d* measure of information gain from after seeing the data in
Bayesian inference
}
