% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstar.R
\name{probability_comparison}
\alias{probability_comparison}
\title{Compares probability densities across one reference model to another}
\usage{
probability_comparison(prob1, prob2)
}
\arguments{
\item{prob1}{Set of probability density values from one model (this is reference set)}

\item{prob2}{Set of probability density values from another model (this is comparison set)}
}
\value{
the fraction of cases where prob1 > prob2
}
\description{
Compares probability densities across one reference model to another
}
