% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstar.R
\name{ash2}
\alias{ash2}
\title{Replicates ash1 but without print statement}
\usage{
ash2(bins)
}
\arguments{
\item{bins}{number of bins}
}
\value{
list which is useable by kde methods
}
\description{
Replicates ash1 but without print statement
}
